Redid database methods
Unfortuently the previous commit broke some of my project and had to be reverted. Work was lost but I recoded most of the work I had done and figured out what needed to be done.

Redid Serialization of objects and fixed salting and hashing. I should have remade all of the database progress i had made this weekend